Rename realm to storage and widen the pin pipeline contracts
src/realm/realm.ts becomes src/storage/storage.ts - it has been the
AsyncStorage wrapper since the Realm migration and the old name only
misleads. All 18 importers and CLAUDE.md updated.

decryptKey accepts string | null | undefined for data and key: any
failure inside is caught and yields undefined, which every caller
already handles, so the strict params were rejecting values the
pipeline genuinely carries (Config envs, optional stored tokens).
getDigitPinCode and refreshSCToken widen to match. Storage catch
blocks are typed any so the legacy return-the-error pattern keeps its
loose shape for callers.
