+++ src/screens/settings/container/settingsContainer.tsx
@@ -31,7 +31,7 @@ import {
   setAuthStatus,
   setExistUser,
   removeAllUserData,
-} from '../../../realm/realm';
+} from '../../../storage/storage';
 
 // Services and Actions
 import {
